+++ app/upload/page.tsx
@@ -4,7 +4,7 @@ import { Fragment, useEffect, useState } from 'react';
 import { useFieldArray, useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { Plus, X } from 'lucide-react';
-import FullScreenLoading from '~/components/Loading/FullScreenLoading';
+import FullScreenLoading from '~/components/common/Loading/FullScreenLoading';
 import SelectForm from '~/components/common/SelectForm/SelectForm';
 import { Button } from '~/components/ui/button';
 import {
